# openapi.yaml
openapi: 3.1.0
info:
  title: 6S Issue Management & Visual Reporting System API
  version: 1.0.0
  description: API phục vụ ghi nhận, đồng bộ offline, giải quyết và tính điểm sức khỏe 6S trong môi trường nhà xưởng.
servers:
  - url: /api
    description: Runtime-configured API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TranslationContext:
      type: object
      properties:
        category:
          type: string
        cause_type:
          type: string
        location_code:
          type: string
        location_name:
          type: string
        tags:
          type: array
          items:
            type: string
...
security:
  - BearerAuth: []
paths:
...
  /ai/cached:
    post:
      operationId: getCachedTranslation
      summary: Tra cache dịch trước khi gọi provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, target_lang]
              properties:
                text:
                  type: string
                target_lang:
                  type: string
                context:
                  $ref: '#/components/schemas/TranslationContext'
      responses:
        '200':
          description: Kết quả cache
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      cached:
                        type: boolean
                      translated_text:
                        type: string
...